# domains/core-platform-services/auth-service/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/academy/teaching-schedules/course/{courseId}": {
      "get": {
        "description": "Accessible by enrolled users, instructors, or admin",
        "operationId": "TeachingScheduleController_getCourseSchedules",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all schedules for a specific course",
        "tags": [
          "Teaching Schedules"
        ]
      }
    },
    "/academy/teaching-schedules": {
      "post": {
        "description": "🔒 Instructor / Admin only",
        "operationId": "TeachingScheduleController_createSchedule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeachingScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a teaching schedule",
        "tags": [
          "Teaching Schedules"
        ]
      }
    },
    "/academy/teaching-schedules/{id}": {
      "put": {
        "description": "🔒 Instructor / Admin only",
        "operationId": "TeachingScheduleController_updateSchedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeachingScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a teaching schedule",
        "tags": [
          "Teaching Schedules"
        ]
      },
      "delete": {
        "description": "🔒 Instructor / Admin only",
        "operationId": "TeachingScheduleController_deleteSchedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a teaching schedule",
        "tags": [
          "Teaching Schedules"
        ]
      },
      "get": {
        "description": "Accessible by instructor/admin or enrolled users",
        "operationId": "TeachingScheduleController_getScheduleById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a teaching schedule by ID",
        "tags": [
          "Teaching Schedules"
        ]
      }
    },
...
    "/contech/client-reports": {
      "post": {
        "operationId": "ClientReportController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClientReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a new client report",
        "tags": [
          "Client Reports"
        ]
      }
    },
    "/contech/client-reports/projects/{projectId}": {
      "get": {
        "operationId": "ClientReportController_findAllByProjectId",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID of the project",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all client reports for a project",
        "tags": [
          "Client Reports"
        ]
      }
    },
    "/contech/client-reports/{id}": {
      "get": {
        "operationId": "ClientReportController_findOneById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID of the client report",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a client report by its ID",
        "tags": [
          "Client Reports"
        ]
      }
    },
...
    "/careers/jobs": {
      "post": {
        "operationId": "CareersController_createJob",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Careers"
        ]
      },
      "get": {
        "operationId": "CareersController_findAllJobs",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Careers"
        ]
      }
    },
...
    "/careers/jobs/{id}": {
      "get": {
        "operationId": "CareersController_findJobById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Careers"
        ]
      },
      "put": {
        "operationId": "CareersController_updateJob",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Careers"
        ]
      },
      "delete": {
        "operationId": "CareersController_removeJob",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Careers"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Gateway",
    "description": "Gateway documentation for frontend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTeachingScheduleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the schedule",
            "example": "NestJS Live Class"
          },
          "description": {
            "type": "string",
            "description": "Optional description of the schedule",
            "example": "Deep dive into NestJS modules"
          },
          "startTime": {
            "type": "string",
            "description": "Start time of the schedule",
            "example": "2025-12-20T10:00:00Z"
          },
          "endTime": {
            "type": "string",
            "description": "End time of the schedule",
            "example": "2025-12-20T11:30:00Z"
          },
          "type": {
            "type": "string",
            "description": "Type of schedule",
            "enum": [
              "LIVE",
              "RECORDING",
              "Q_AND_A",
              "OFFICE_HOURS",
              "WORKSHOP"
            ],
            "example": "LIVE"
          },
          "courseId": {
            "type": "number",
            "description": "Associated course ID",
            "example": 1
          },
          "isRecurring": {
            "type": "boolean",
            "description": "Whether the schedule is recurring",
            "example": false
          },
          "recurrencePattern": {
            "type": "string",
            "description": "Recurrence pattern (if recurring)",
            "example": "WEEKLY"
          }
        },
        "required": [
          "title",
          "startTime",
          "endTime",
          "type",
          "courseId"
        ]
      },
      "UpdateTeachingScheduleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Updated title",
            "example": "Updated Live Class"
          },
          "description": {
            "type": "string",
            "description": "Updated description",
            "example": "Updated description content"
          },
          "startTime": {
            "type": "string",
            "description": "Updated start time",
            "example": "2025-12-21T10:00:00Z"
          },
          "endTime": {
            "type": "string",
            "description": "Updated end time",
            "example": "2025-12-21T11:30:00Z"
          },
          "type": {
            "type": "string",
            "description": "Updated type of schedule",
            "enum": [
              "LIVE",
              "RECORDING",
              "Q_AND_A",
              "OFFICE_HOURS",
              "WORKSHOP"
            ]
          },
          "isRecurring": {
            "type": "boolean",
            "description": "Updated recurring status",
            "example": true
          },
          "recurrencePattern": {
            "type": "string",
            "description": "Updated recurrence pattern",
            "example": "DAILY"
          }
        }
      },
...
      "KpiDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the KPI",
            "example": "Budget Variance",
            "maxLength": 100
          },
          "value": {
            "type": "string",
            "description": "Current value of the KPI",
            "example": "+5%",
            "maxLength": 50
          },
          "target": {
            "type": "string",
            "description": "Target for the KPI",
            "example": "<2%",
            "maxLength": 50
          }
        },
        "required": [
          "name",
          "value",
          "target"
        ]
      },
      "CreateClientReportDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the client report",
            "example": "Q1 2025 Project Report"
          },
          "projectId": {
            "type": "number",
            "description": "ID of the project",
            "example": 123
          },
          "summary": {
            "type": "string",
            "description": "Summary of the report",
            "example": "This report provides an overview of the project status, including key achievements and challenges.",
            "minLength": 20,
            "maxLength": 5000
          },
          "KPIs": {
            "description": "List of KPIs for the project",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KpiDto"
            }
          }
        },
        "required": [
          "title",
          "projectId",
          "summary",
          "KPIs"
        ]
      },
...
      "CreateJobDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateJobDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}